*change: [KRF-00001] changed project structure pt4 logic
