+++ src/components/common/presentation/display/Table.tsx
@@ -3,7 +3,7 @@ import Row from "../../structure/Row";
 import Column from "../../structure/Column";
 import PageSelector from "../../inputs/PageSelector";
 import WithTooltip, { ToolTipColor, ToolTipPosition } from "../wrapper/WithTooltip";
-import useTranslation from "../../../../logic/context/pageText/getTranslation";
+import useTranslation from "../../../../logic/functions/getTranslation";
 import { IDictionary, PageType } from "../../../../logic/functions/misc";
 
 export interface ITableCell {
