+++ src/components/main/MenuComponents/SubMenuMobile.tsx
@@ -3,7 +3,7 @@ import { ISubMenuItem } from './SubMenu';
 import Row from '../../common/structure/Row';
 import Column from '../../common/structure/Column';
 import PageSelector from '../../common/inputs/PageSelector';
-import useTranslation from '../../../logic/context/pageText/getTranslation';
+import useTranslation from '../../../logic/functions/getTranslation';
 import { AppContext, ErrorContext, LoginContext, RouteContext } from '../../../logic/config/AppProvider';
 
 const SubMenuMobile: React.FC<{ SubMenus: ISubMenuItem[], collapseFunc: () => void }> = ( props ) => {
