+++ src/logic/context/App/appContextEnums.ts
