+++ src/logic/context/Error/appErrorEnums.ts
