+++ src/logic/context/Error/appErrorInterfaces.ts
@@ -1,4 +1,18 @@
-import { ErrorCodes } from "../appErrorEnums";
+import { ErrorActions, ErrorCodes } from "./appErrorEnums";
+
+export interface IError {
+    hasError: boolean;
+    errorDescription?: string;
+    errorCode?: ErrorCodes;
+}
+
+export interface IErrorAction {
+    type: ErrorActions;
+    errorCode?: ErrorCodes;
+    errorDescription?: string;
+}
+
+export type ErrorContextType = [ IError, ( ( action: IErrorAction ) => void ) ];
 
 export interface IErrorData {
     Title: string;
