--- src/logic/context/pageText/pageTranslationInterfaces.ts
@@ -1,5 +1,0 @@
-export interface ITranslation {
-    [ process: string ]: {
-        [ token: string ]: string;
-    };
-}
